TTRequestLoader fix.
TTRequestLoader will fail to correctly create an NSURLRequest when more than one
request has been added to its _requests array, opting instead to pass "nil" to
[TTURLRequestQueue createNSURLRequest:URL:], which results in a bevy of problems
when it comes time to dispatch the request (i.e., the HTTP body / HTTP method
aren't set properly). Since there's no point in initiating the request if it's
already running, I added a check here to see if the request is loading, and I
make sure to always pass a non-nil request (assuming _requests isn't empty,
which it should never be).

With this fix in place, we have to be careful that two different requests with
different parameters (different HTTP methods, different bodies, etc.) never get
added to the same loader, but that shouldn't ever happen since requests are
added to loaders contingent on the output of -[TTURLRequest generateCacheKey],
and that method takes into account the params we're interested in.